moved tests to enricher
